Update docs and move optimization registration code
