+++ core/src/main/scala/com/carto/analyticstoolbox/spark/sql/SpatialFilterPushdownOptimizations.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.carto.analyticstoolbox.spark
+package com.carto.analyticstoolbox.spark.sql
 
 import org.apache.spark.sql.SparkSessionExtensions
 
@@ -22,5 +22,5 @@ import com.carto.analyticstoolbox.spark.rules.SpatialFilterPushdownRules
 
 class SpatialFilterPushdownOptimizations extends (SparkSessionExtensions => Unit) {
   def apply(e: SparkSessionExtensions): Unit =
-    e.injectOptimizerRule(spark => SpatialFilterPushdownRules)
+    e.injectOptimizerRule(_ => SpatialFilterPushdownRules)
 }
